% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyglotSQL-package.R
\docType{package}
\name{polyglotSQL-package}
\alias{polyglotSQL}
\alias{polyglotSQL-package}
\title{polyglotSQL: SQL Parsing, Analysis and Dialect Translation}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Parse, tokenize, validate, format, analyze and translate SQL between more than 30 dialects ('PostgreSQL', 'MySQL', 'BigQuery', 'Snowflake', 'DuckDB', 'T-SQL', and others) using the 'polyglot-sql' Rust crate \url{https://github.com/tobilg/polyglot}, a Rust port of the 'SQLGlot' 'Python' library. All processing happens natively in process; no database connection, 'Python' runtime or external service is required. Includes column-level lineage, structural query analysis, query optimization, 'AST' diffing and 'OpenLineage' facet generation.
}
\section{Acknowledgements}{

polyglotSQL embeds the \href{https://github.com/tobilg/polyglot}{polyglot-sql}
Rust crate by Tobias Müller (MIT), which is a Rust port of
\href{https://github.com/tobymao/sqlglot}{SQLGlot} by Toby Mao (MIT). See
\code{inst/COPYRIGHTS} for the licenses of all vendored Rust dependencies.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/polyglot-sql-r}
  \item \url{https://strategicprojects.github.io/polyglot-sql-r/}
  \item Report bugs at \url{https://github.com/StrategicProjects/polyglot-sql-r/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@de.ufpe.br}

Authors:
\itemize{
  \item André Leite \email{leite@de.ufpe.br}
  \item Tobias Müller \email{github@tobilg.com} (Author of the bundled 'polyglot-sql' Rust crate (Polyglot project)) [copyright holder]
  \item Toby Mao (Author of SQLGlot, from which Polyglot is derived) [copyright holder]
}

Other contributors:
\itemize{
  \item  The authors of the vendored Rust dependencies (see inst/COPYRIGHTS) [copyright holder]
}

}
\keyword{internal}
